Create separarte parsers from each bank
