+++ rules_manager.py
